requirements_test.txt
aniso8601==5.1.0
Django==2.2.2
django-filter==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.9.4
djangorestframework-filters==1.0.0.dev0
djangorestframework-jwt==1.11.0
djoser==1.7.0
graphene==2.1.5
graphene-django==2.3.0
graphql-core==2.2
graphql-relay==0.4.5
promise==2.2.1
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
relativetimebuilder==1.0.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
